code formatted

+++ zombiehelper.sol
@@ -4,8 +4,8 @@ pragma solidity <0.9.0;
 import "./ZombieFeeding.sol";
 
 contract ZombieHelper is ZombieFeeding {
-
-    uint levelUpFee = 0.001 ether;
+    
+    uint256 levelUpFee = 0.001 ether;
 
     modifier aboveLevel(uint256 _level, uint256 _zombieId) {
         require(zombies[_zombieId].level >= _level);
@@ -16,11 +16,11 @@ contract ZombieHelper is ZombieFeeding {
         payable(owner).transfer(address(this).balance);
     }
 
-    function setLevelUpFee(uint _fee) external onlyOwner {
+    function setLevelUpFee(uint256 _fee) external onlyOwner {
         levelUpFee = _fee;
     }
 
-    function levelUp(uint _zombieId) external payable {
+    function levelUp(uint256 _zombieId) external payable {
         require(msg.value == levelUpFee);
         zombies[_zombieId].level++;
     }
@@ -38,20 +38,23 @@ contract ZombieHelper is ZombieFeeding {
         aboveLevel(20, _zombieId)
     {
         require(msg.sender == zombieToOwner[_zombieId]);
-      	zombies[_zombieId].dna = _newDna;
+        zombies[_zombieId].dna = _newDna;
     }
 
-	function getZombiesByOwner(address _owner) external view returns (uint[] memory) {
-		uint[] memory result = new uint[](ownerZombieCount[_owner]);
+    function getZombiesByOwner(address _owner)
+        external
+        view
+        returns (uint256[] memory)
+    {
+        uint256[] memory result = new uint256[](ownerZombieCount[_owner]);
 
-		uint counter = 0;
-		for (uint i = 0; i < zombies.length; i++) {
-			if (zombieToOwner[i] == _owner) {
-				result[counter] = i;
-				counter++;
-			}
-		}
-		return result;
-
-	}
+        uint256 counter = 0;
+        for (uint256 i = 0; i < zombies.length; i++) {
+            if (zombieToOwner[i] == _owner) {
+                result[counter] = i;
+                counter++;
+            }
+        }
+        return result;
+    }
 }
